Audio Syncing solution merged and imported from modified alt branch
- Audio Syncing solution merged and imported from modified alt branch
- New Scene created for this, Game_AudioSync
- New Music by Jeremey added
- Some Folder structures sorted

We have a problem with using prefabs, if one scene disables them or does
something else it will be shown/modified in the pre-fab in another scene
which uses them. This is what will happen if you try to play
game_audiosync at the moment. Nothing is broken, the prefabs are just
getting disabled so they don't show in when you go back to main. Just
re-enable them.

There are hard couplings in code which I'm not sure how to deal with
100% cleanly yet.